requirements.txt
Adafruit-BBIO==0.0.30
https://github.com/philipbl/CoAPy/archive/python3.zip
msgpack-python==0.4.8
paho-mqtt==1.2
pyserial==3.1.1
python-persistent-queue==1.3.0
pytz==2016.10
tabulate==0.7.7
